similarity/requirements.txt
torch
sentence-transformers
fugashi
ipadic
unidic-lite
pandas
